+++ docs/CHANGELOG.md
@@ -5,6 +5,9 @@
 * Add TerminalApi - `searchTamperedTerminal` API
 * Add parameter query `includeGroupInfo` and return `groups` in TerminalApi - `getTerminal`, `getTerminalBySn` API
 
+### Improvement
+* Make fields `contact`, `country` and `phone` not mandatory in ResellerApi - `createReseller`, `updateReseller` API
+
 
 ## 11.0.1
 ### Bug Fix